Estimate Genetic Map Positions

Designed to work with one chromosome at a time. estimate_map calculates genetic map positions using recombination frequency (RF) and LOD scores calculated from a genotype matrix. It applies a selected mapping function (Kosambi or Haldane) and outputs map positions (in centi Morgans, cM) at multiple resolutions defined by the user.

Usage

estimate_map(
  geno_matrix,
  model = "Kosambi",
  n_points = seq(10, 20, by = 5),
  pop.type = "F2"
)

Arguments

geno_matrix

A genotype matrix where rows represent markers and columns represent individuals. This matrix must be in a format compatible with MapRtools::MLEL and with markers for only one chromosome or Linkage Group.

model

Character string specifying the mapping function to use. Options are "Kosambi" (default) or "Haldane".

n_points

A numeric vector indicating the different values of n.point to be used for estimating genetic positions. Default is seq(10, 20, by = 5).

pop.type

Character string specifying the population type. Supported options are "DH", "BC", "F2" (default), "S1", "RIL.self", and "RIL.sib".

Value

A data frame containing the extracted genetic map with additional columns (pN) corresponding to the estimated positions for each specified n.point value.

Details

The function uses MapRtools::MLEL() to compute RF and LOD matrices and then applies the specified mapping function via MapRtools::map_fn(). For each n.point value, the function appends a new column (pN) to the map with the estimated genetic positions.

Examples

if (FALSE) { # \dontrun{
# Example genotype matrix

geno_matrix <- chr1_geno_matrix

map <- estimate_map(chr1_geno_matrix, model = "Kosambi", n_points = c(10, 15, 20), pop.type = "F2")

# Simply
genetic_map_chr1 <- estimate_map(chr1_geno_matrix)
} # }