Get LOD Support Intervals for QTL Peaks (Single or Multiple Traits)

Computes LOD support intervals for detected QTLs from scanone or cim.

Usage

get_LOD_int(cross_obj, model_obj, results_obj, trait = NULL, drop = 1.5)

Arguments

cross_obj

A cross object from the qtl package. Usually the output of read.cross or calc.genoprob.

model_obj

Either a scanone/CIM object (class "scanone") or a named list of such objects. Usuall the output of cim or scanone from qtl.

results_obj

Either a data frame of QTL results (with "chr" and "pos") or a named list of such data frames.

trait

Optional. Name of the trait to analyze. Required if model_obj and results_obj are lists (i.e., when working with multiple traits).

drop

Numeric. LOD drop to define the support interval (default = 1.5).

Value

A named list of QTL interval summaries, flanking markers, peak marker (pseudo marker or physical marker), and usable marker (closest physical marker if peak is on a pseudo marker).

Details

Works with either a single model and qtl results or a list of models and results for multiple traits.

Examples

if (FALSE) { # \dontrun{
# Multiple traits:
# For the length-width trait
lw_result_pop1 <- get_LOD_int(
  cross_obj = M1,
  model_obj = cim_qtl_results1,
  results_obj = results_pop1,
  trait = "length_width_ratio",
  drop = 1.5
)

# Single traits:
mod <- cim_qtl_results1[["length_width_ratio"]]
res <- results_pop1[["length_width_ratio"]]

single <- get_LOD_int(
  cross_obj = M1,
  model_obj = mod,
  results_obj = res
)
} # }